Implementing SQLite as an alternative to MySQL
